メソッドや型定義の配置整理
