+++ lib/misskey/misskey.go
@@ -1,22 +1,23 @@
 package misskey
 
 import (
-	"bytes"
-	"context"
 	"encoding/json"
 	"fmt"
 	"hato-bot-go/lib/amesh"
 	"io"
-	"log"
-	"mime/multipart"
 	"net/http"
-	"os"
 	"strings"
 	"time"
 
 	"github.com/cockroachdb/errors"
-	"github.com/gorilla/websocket"
 )
+
+// BotSetting Misskeyボットの設定
+type BotSetting struct {
+	Domain string       // Misskeyのドメイン
+	Token  string       // APIトークン
+	Client *http.Client // HTTPクライアント
+}
 
 // Note Misskeyのノート構造体
 type Note struct {
@@ -33,10 +34,11 @@ type Note struct {
 	} `json:"user"`
 }
 
-// ParseResult ameshコマンドの解析結果を表す構造体
-type ParseResult struct {
-	Place   string
-	IsAmesh bool
+// CreateNoteRequest ノート作成のリクエスト構造体
+type CreateNoteRequest struct {
+	Text         string   // ノートのテキスト
+	FileIDs      []string // 添付ファイルのID一覧
+	OriginalNote *Note    // 返信元のノート
 }
 
 // File アップロードされたファイルの構造体
@@ -46,34 +48,10 @@ type File struct {
 	URL  string `json:"url"`
 }
 
-// CreateNoteRequest ノート作成のリクエスト構造体
-type CreateNoteRequest struct {
-	Text         string   // ノートのテキスト
-	FileIDs      []string // 添付ファイルのID一覧
-	OriginalNote *Note    // 返信元のノート
-}
-
-// BotSetting Misskeyボットの設定
-type BotSetting struct {
-	Domain string       // Misskeyのドメイン
-	Token  string       // APIトークン
-	Client *http.Client // HTTPクライアント
-}
-
-// Bot Misskeyボットクライアント
-type Bot struct {
-	BotSetting *BotSetting
-	UserAgent  string
-	WSConn     *websocket.Conn
-}
-
-// NewBot 新しいBotインスタンスを作成
-func NewBot(domain, token string) *Bot {
-	return NewBotWithClient(&BotSetting{
-		Domain: domain,
-		Token:  token,
-		Client: &http.Client{Timeout: 30 * time.Second},
-	})
+// ParseResult ameshコマンドの解析結果を表す構造体
+type ParseResult struct {
+	Place   string
+	IsAmesh bool
 }
 
 // NewBotWithClient HTTPクライアント注入可能なBotインスタンスを作成
@@ -90,290 +68,13 @@ func NewBotWithClient(botSetting *BotSetting) *Bot {
 	}
 }
 
-// Connect WebSocket接続を確立
-func (bot *Bot) Connect() error {
-	wsURL := fmt.Sprintf("wss://%s/streaming?i=%s", bot.BotSetting.Domain, bot.BotSetting.Token)
-
-	dialer := websocket.DefaultDialer
-	dialer.HandshakeTimeout = 10 * time.Second
-
-	conn, _, err := dialer.Dial(wsURL, http.Header{
-		"User-Agent": []string{bot.UserAgent},
+// NewBot 新しいBotインスタンスを作成
+func NewBot(domain, token string) *Bot {
+	return NewBotWithClient(&BotSetting{
+		Domain: domain,
+		Token:  token,
+		Client: &http.Client{Timeout: 30 * time.Second},
 	})
-	if err != nil {
-		return errors.Wrap(err, "Failed to Dial")
-	}
-
-	bot.WSConn = conn
-
-	// メインチャンネルに接続
-	connectMsg := struct {
-		Type string      `json:"type"`
-		Body interface{} `json:"body,omitempty"`
-	}{
-		Type: "connect",
-		Body: map[string]interface{}{
-			"channel": "main",
-			"id":      "main",
-		},
-	}
-
-	if err := bot.WSConn.WriteJSON(connectMsg); err != nil {
-		return errors.Wrap(err, "Failed to WriteJSON")
-	}
-
-	log.Printf("Connected to Misskey WebSocket: %s", bot.BotSetting.Domain)
-	return nil
-}
-
-// Listen WebSocketメッセージを監視
-func (bot *Bot) Listen(messageHandler func(note *Note)) error {
-	if messageHandler == nil {
-		return errors.New("messageHandler cannot be nil")
-	}
-
-	for {
-		var msg struct {
-			Type string `json:"type"`
-			Body struct {
-				ID   string `json:"id"`
-				Type string `json:"type"`
-				Body Note   `json:"body"`
-			} `json:"body"`
-		}
-		if err := bot.WSConn.ReadJSON(&msg); err != nil {
-			return errors.Wrap(err, "Failed to ReadJSON")
-		}
-
-		// メンションイベントの処理
-		if msg.Type != "channel" || msg.Body.Type != "mention" {
-			continue
-		}
-
-		note := msg.Body.Body
-		log.Printf("Received mention from @%s: %s", note.User.Username, note.Text)
-
-		// メッセージハンドラーを呼び出し
-		messageHandler(&note)
-	}
-}
-
-// CreateNote ノートを作成
-func (bot *Bot) CreateNote(ctx context.Context, req *CreateNoteRequest) error {
-	if req == nil {
-		return errors.New("req cannot be nil")
-	}
-	if req.OriginalNote == nil {
-		return errors.New("originalNote cannot be nil")
-	}
-
-	// noteから必要な情報を取得
-	visibility := req.OriginalNote.Visibility
-	replyID := req.OriginalNote.ID
-
-	// 公開範囲がpublicならばhomeにする
-	if visibility == "public" {
-		visibility = "home"
-	}
-
-	data := map[string]interface{}{
-		"text":       req.Text,
-		"visibility": visibility,
-	}
-
-	if replyID != "" {
-		data["replyId"] = replyID
-	}
-
-	if 0 < len(req.FileIDs) {
-		data["fileIds"] = req.FileIDs
-	}
-
-	// 元の投稿がCWされていた場合、それに合わせてCW投稿する
-	if req.OriginalNote.CW != nil {
-		data["cw"] = "隠すっぽ！"
-	}
-
-	resp, err := bot.apiRequest(ctx, "notes/create", data)
-	if err != nil {
-		return errors.Wrap(err, "Failed to apiRequest")
-	}
-
-	var result struct {
-		CreatedNote Note `json:"createdNote"`
-	}
-
-	if err := checkStatusAndDecodeJSON(resp, &result); err != nil {
-		return errors.Wrap(err, "Failed to checkStatusAndDecodeJSON")
-	}
-
-	return nil
-}
-
-// UploadFileFromReader io.Readerからファイルをアップロード
-func (bot *Bot) UploadFileFromReader(ctx context.Context, reader io.Reader, fileName string) (*File, error) {
-	var buf bytes.Buffer
-	writer := multipart.NewWriter(&buf)
-
-	// トークンフィールドを追加
-	if writeErr := writer.WriteField("i", bot.BotSetting.Token); writeErr != nil {
-		return nil, errors.Wrap(writeErr, "Failed to WriteField")
-	}
-
-	// ファイルフィールドを追加
-	part, err := writer.CreateFormFile("file", fileName)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to CreateFormFile")
-	}
-
-	if _, copyErr := io.Copy(part, reader); copyErr != nil {
-		return nil, errors.Wrap(copyErr, "Failed to io.Copy")
-	}
-
-	if closeErr := writer.Close(); closeErr != nil {
-		return nil, errors.Wrap(closeErr, "Failed to Close")
-	}
-
-	url := fmt.Sprintf("https://%s/api/drive/files/create", bot.BotSetting.Domain)
-	req, err := http.NewRequestWithContext(ctx, "POST", url, &buf)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to http.NewRequestWithContext")
-	}
-
-	req.Header.Set("Content-Type", writer.FormDataContentType())
-	req.Header.Set("User-Agent", bot.UserAgent)
-
-	resp, err := bot.BotSetting.Client.Do(req)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to Do")
-	}
-
-	var uploadedFile File
-	if err := checkStatusAndDecodeJSON(resp, &uploadedFile); err != nil {
-		return nil, errors.Wrap(err, "Failed to checkStatusAndDecodeJSON")
-	}
-
-	return &uploadedFile, nil
-}
-
-// AddReaction リアクションを追加
-func (bot *Bot) AddReaction(ctx context.Context, noteID, reaction string) (err error) {
-	data := map[string]interface{}{
-		"noteId":   noteID,
-		"reaction": reaction,
-	}
-
-	resp, err := bot.apiRequest(ctx, "notes/reactions/create", data)
-	if err != nil {
-		return errors.Wrap(err, "Failed to apiRequest")
-	}
-	defer func(body io.ReadCloser) {
-		if closeErr := body.Close(); closeErr != nil {
-			err = errors.Wrap(closeErr, "Failed to Close")
-		}
-	}(resp.Body)
-
-	if resp.StatusCode != 204 {
-		return fmt.Errorf("API returned status %d", resp.StatusCode)
-	}
-
-	return nil
-}
-
-// ProcessAmeshCommand ameshコマンドを処理
-func (bot *Bot) ProcessAmeshCommand(ctx context.Context, note *Note, place string) error {
-	if note == nil {
-		return errors.New("note cannot be nil")
-	}
-
-	// 処理中リアクションを追加
-	if err := bot.AddReaction(ctx, note.ID, "👀"); err != nil {
-		return errors.Wrap(err, "Failed to AddReaction")
-	}
-
-	// Yahoo APIキーを取得
-	apiKey := os.Getenv("YAHOO_API_TOKEN")
-	if apiKey == "" {
-		return errors.New("YAHOO_API_TOKEN environment variable not set")
-	}
-
-	// 位置を解析
-	location, err := amesh.ParseLocation(place, apiKey)
-	if err != nil {
-		return errors.Wrap(err, "Failed to amesh.ParseLocation")
-	}
-
-	fmt.Printf("Generating amesh image for %s (%.4f, %.4f)\n", location.PlaceName, location.Lat, location.Lng)
-
-	// 画像をメモリ上に作成
-	imageReader, err := amesh.CreateImageReader(ctx, location)
-	if err != nil {
-		return errors.Wrap(err, "Failed to amesh.CreateImageReader")
-	}
-
-	// ファイル名を生成
-	fileName := amesh.GenerateFileName(location)
-
-	// Misskeyにメモリから直接アップロード
-	uploadedFile, err := bot.UploadFileFromReader(ctx, imageReader, fileName)
-	if err != nil {
-		return errors.Wrap(err, "Failed to UploadFileFromReader")
-	}
-
-	// 結果をノートとして投稿
-	text := fmt.Sprintf(
-		"📡 %s (%.4f, %.4f) の雨雲レーダー画像だっぽ",
-		location.PlaceName,
-		location.Lat,
-		location.Lng,
-	)
-	if err := bot.CreateNote(ctx, &CreateNoteRequest{
-		Text:         text,
-		FileIDs:      []string{uploadedFile.ID},
-		OriginalNote: note,
-	}); err != nil {
-		return errors.Wrap(err, "Failed to CreateNote")
-	}
-
-	log.Printf("Successfully processed amesh command for %s", location.PlaceName)
-	return nil
-}
-
-// apiRequest MisskeyAPIリクエストを送信
-func (bot *Bot) apiRequest(ctx context.Context, endpoint string, data interface{}) (*http.Response, error) {
-	// データにトークンを追加
-	payload := map[string]interface{}{
-		"i": bot.BotSetting.Token,
-	}
-
-	if data != nil {
-		if dataMap, ok := data.(map[string]interface{}); ok {
-			for k, v := range dataMap {
-				payload[k] = v
-			}
-		}
-	}
-
-	jsonData, err := json.Marshal(payload)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to json.Marshal")
-	}
-
-	url := fmt.Sprintf("https://%s/api/%s", bot.BotSetting.Domain, endpoint)
-	req, err := http.NewRequestWithContext(ctx, "POST", url, bytes.NewBuffer(jsonData))
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to http.NewRequestWithContext")
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("User-Agent", bot.UserAgent)
-
-	resp, err := bot.BotSetting.Client.Do(req)
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to Do")
-	}
-
-	return resp, nil
 }
 
 // ParseAmeshCommand ameshコマンドを解析
